refactor(train): S4b-① —— models.py→vae.py 名实相符 + ctx 文本编码器字段收敛 text_stack（多模型 PR-2b，D15/拍板项3）
- training/models.py 瘦身后仅剩 VAE 内容，git mv 为 training/vae.py；
  ensure_models_namespace 迁 model_loading.py（路径工具的家）
- TrainingContext.qwen_model/qwen_tok/t5_tok 三字段退役 → text_stack
  （family 私有结构，对循环 opaque）
