+++ tests/test_training_sample_runner.py
@@ -48,9 +48,7 @@ def _ctx(*, sample_negative_prompt: str = "") -> SimpleNamespace:
         args=args,
         model=torch.nn.Linear(1, 1),
         vae=object(),
-        qwen_model=object(),
-        qwen_tok=object(),
-        t5_tok=object(),
+        text_stack=(object(), object(), object()),
         optimizer=object(),
         injector=FakeInjector(),
         device="cpu",
